scripts/requirements.txt
# scripts/requirements.txt
pandas==2.0.3
scikit-learn==1.3.2
numpy==1.26.4
ollama>=0.5.1
python-dotenv==1.0.1